requirements.txt
numpy>=1.24.0,<2.0  # UHD requiere NumPy 1.x
scipy>=1.7.0
matplotlib>=3.4.0
pyyaml>=5.4.0
h5py>=3.0.0  # Para leer archivos .mat en formato MATLAB v7.3 (HDF5)
py3gpp

# NOTA: uhd no está disponible en PyPI
# Instalación en Ubuntu/Debian:
#   sudo apt install python3-uhd uhd-host libuhd-dev
# Luego crear enlace simbólico al entorno virtual:
#   ln -s /usr/lib/python3/dist-packages/uhd env/lib/python3.12/site-packages/uhd
